Improve statuses
- The former `reasonPhrase` of `Response` is now the computed property
`description` of `Status`
- Custom statuses can now be initialized with a custom description,
which will then be returned by the `description` property
- Rather than always setting the status to 200, the convenience `send`
methods have the additional `status` parameter now. It has a default
value (200), so this change won’t effect existing code, but allows
setting any status in future.
- The required `send` method accepts an optional `status` parameter
now. This way, the status can be set by passing it to the method, but
at the same time won’t overwrite a previously set status if not needed.
- When an error occurs in one of the convenience `send` methods, they
will now call the `send(error:)` method and return.